feat: move 5 more zero-dep files to commons (batch 7)
Move TorServiceStatus, ShorterTopAppBar, FilterPostsByScopes, FeedBasis,
and FilterCommunityPosts to commons/commonMain.

Files moved:
- TorServiceStatus.kt (sealed class, no deps)
- ShorterTopAppBar.kt + TopBarSize (pure Compose Material3)
- FilterPostsByScopes.kt (quartz-only deps)
- FeedBasis.kt (quartz-only deps)
- FilterCommunityPosts.kt (quartz-only deps)

Build verified: commons:compileKotlinJvm + amethyst:compilePlayDebugKotlin pass.
