+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/tor/TorServiceStatus.kt
@@ -18,32 +18,14 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.screen.loggedIn.home.datasource.nip22Comments
+package com.vitorpamplona.amethyst.commons.ui.tor
 
-import com.vitorpamplona.quartz.nip01Core.relay.client.pool.RelayBasedFilter
-import com.vitorpamplona.quartz.nip01Core.relay.filters.Filter
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip22Comments.CommentEvent
+sealed class TorServiceStatus {
+    data class Active(
+        val port: Int,
+    ) : TorServiceStatus()
 
-val CommentKinds = listOf(CommentEvent.KIND)
+    object Off : TorServiceStatus()
 
-fun filterHomePostsByScopes(
-    relay: NormalizedRelayUrl,
-    scopesToLoad: Set<String>,
-    since: Long?,
-): List<RelayBasedFilter> {
-    if (scopesToLoad.isEmpty()) return emptyList()
-
-    return listOf(
-        RelayBasedFilter(
-            relay = relay,
-            filter =
-                Filter(
-                    kinds = CommentKinds,
-                    tags = mapOf("I" to scopesToLoad.toList()),
-                    limit = 100,
-                    since = since,
-                ),
-        ),
-    )
+    object Connecting : TorServiceStatus()
 }
